# streamex-api-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Movie details with streaming sources.\nNote: 'episodes' for a movie will typically contain a single entry representing the full movie.",
        "contact": {}
    },
    "paths": {
        "/api/detail/{category}/{id}": {
            "get": {
                "description": "Mengambil detail lengkap untuk serial TV atau anime, termasuk metadata dan daftar sumber streaming.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Dapatkan detail konten (TV atau Anime)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kategori konten: 'tv' atau 'anime'",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID konten (TMDB ID untuk tv, HiAnime ID untuk anime)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1",
                        "description": "Nomor musim (Hanya berlaku untuk 'tv')",
                        "name": "season",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Untuk kategori 'tv' atau 'anime'",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/list/{category}": {
            "get": {
                "description": "Menampilkan daftar film, serial TV, atau anime populer. Untuk anime, gunakan parameter 'status'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Daftar konten populer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kategori: tv, movie, anime",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1",
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "most-popular",
                        "description": "Status/Filter khusus Anime: trending, top-airing, most-popular, most-favorite, top-upcoming, completed, top-rated",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/movie/{id}": {
            "get": {
                "description": "Mengambil detail lengkap untuk film, termasuk metadata dan daftar sumber streaming.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Dapatkan detail film",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID film (TMDB ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MovieDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/search": {
            "get": {
                "description": "Mencari film, serial TV, atau anime berdasarkan query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Cari konten",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kata kunci pencarian",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1",
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
}